+++ build.gradle
@@ -22,6 +22,7 @@ subprojects {
 
     sourceCompatibility = '1.8'
     targetCompatibility = '1.8'
+    compileJava.options.encoding = 'UTF-8'
 
     repositories {
         mavenCentral()
@@ -57,27 +58,32 @@ project(':module-common') {
     dependencies {
         implementation 'org.flywaydb:flyway-core'
     }
-
-    configurations {
-        testClasses {
-            extendsFrom(testImplementation)
-        }
-    }
-
-    task testJar(type: Jar) {
-        archiveClassifier.set('test')
-        from sourceSets.test.output
-    }
-
-    // add the jar generated by the testJar task to the testClasses dependency
-    artifacts {
-        testClasses testJar
-    }
 }
 
 project(':module-product') {
+    bootJar {
+        enabled = false
+    }
+    jar {
+        enabled = true
+    }
+
     dependencies {
         implementation project(':module-common')
-        testImplementation project(path: ':module-common', configuration: 'testClasses')
+    }
+}
+
+project(':module-menu') {
+    bootJar {
+        enabled = false
+    }
+    jar {
+        enabled = true
+    }
+
+    dependencies {
+        implementation project(':module-common')
+        implementation project(':module-product')
+        testImplementation project(':module-product')
     }
 }
