+++ module-menu/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+logging.level.org.springframework.jdbc.core=TRACE
+spring.h2.console.enabled=true
+
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.hibernate.type.descriptor.sql=trace
