+++ module-product/src/test/java/kitchenpos/product/ModuleProductApplicationTests.java
@@ -1,9 +1,9 @@
 package kitchenpos.product;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootTest
+@SpringBootApplication
 class ModuleProductApplicationTests {
 
     @Test
