+++ settings.gradle
@@ -1,4 +1,6 @@
 rootProject.name = 'kitchenpos'
 
 include 'module-common'
+include 'module-menu'
 include 'module-product'
+
